+++ src/features_v1/validators/comment_validators.py
@@ -17,6 +17,19 @@ from .problem_validators import (
 
 
 class BaseCommentValidator:
+    """
+    Базовый класс валидаторов для операций с комментариями.
+
+    Назначение:
+        Инкапсулирует часто используемые проверки, связанные с комментариями,
+        такими как проверка принадлежности к треду, проверка владельца комментария,
+        наличие или отсутствие лайка от пользователя.
+
+    Преимущества:
+        - Централизация логики валидации: предотвращает дублирование кода в разных валидаторах.
+        - Повышение читаемости и сопровождаемости: вся логика в одном месте.
+        - Упрощение тестирования и переиспользования логики валидации.
+    """
     def __init__(self, session: AsyncSession):
         self.session = session
